requirements.txt
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
distro==1.9.0
ecdsa==0.19.2
fastapi==0.115.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.13.0
openai==1.51.0
passlib==1.7.4
pyasn1==0.6.3
pycparser==3.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dotenv==1.2.2
python-jose==3.3.0
python-multipart==0.0.12
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tqdm==4.67.3
typing_extensions==4.15.0
uvicorn==0.30.6
watchfiles==1.1.1
websockets==16.0
